Renamed Drop to Dropl
The package actually stands for Deferred Realtime Operations Library and removing the "L" to make it sound like a real word wasn't right move. Additionally, developers may have dilemma picking a variable name that doesn't conflict with a package they have installed, like when working on enemy item drops. Dropl would be non-sense word and would perfectly fit right in as it's not something that has weight of pre-conceived notion.